Private pads description update
Hi,
I just made clear that HTTPS is needed to run Etherpad with Nextcloud and private pads through the API.

I realized that in DisplayController.php line 95, we have setcookie() with the HTTPS only parameter set. I running all locally and I wondering why Etherpad couldn't grant me access to pads, and It was because the session ID was not being sent.

Thank you for your work! :smile: